% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColSummary.R
\name{ColSummary}
\alias{ColSummary}
\title{Minimum, maximum, and variance of discretized kernel column sums used a summary statistics for discretized growth transition kernels}
\usage{
ColSummary(dKernList, index = 1:365)
}
\arguments{
\item{dKernList}{Discretized kernel list from \code{\link{SimulateModel}} [matrix]}

\item{index}{Indices of discretized kernel list objects to include [integer]}
}
\value{
Matrix of min, max, and variance of column sums [matrix]
}
\description{
Calculates Minimum, maximum,and variance of column sums for a list of discretized kernel matrices as output from \code{\link{SimulateModel}}
}
