% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fecundity.R
\name{Fecundity}
\alias{Fecundity}
\title{Kernel Functions - Fecundity}
\usage{
Fecundity(z, pars, date)
}
\arguments{
\item{z}{Size at the beginning of the timestep [float]}

\item{pars}{Data.frame containing the date-indexed parameters[data.frame]}

\item{date}{Ordinal day to reference proper 'pars' date-indexed parameters [integer]}
}
\value{
Number of hatchlings per spawn[binary]
}
\description{
If an individual reaches 'z_repro' they reproduce with probability provided by the \code{\link{Spawning}} function, and have 'hatch_per_spawn' hatchlings
}
\seealso{
Other Kernel Functions: 
\code{\link{HatchlingGrowth}()},
\code{\link{Spawning}()},
\code{\link{Survival}()}
}
\concept{Kernel Functions}
