% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInitialPop.R
\name{GenerateInitialPop}
\alias{GenerateInitialPop}
\title{Generate an Initial Population}
\usage{
GenerateInitialPop(mpts, z_hatch, z_inf, n_0, init_dist = "unif")
}
\arguments{
\item{mpts}{Midpoints of the discretized kernel that the size distribution is being generated for [vector]}

\item{z_hatch}{Minimum size of individuals [float]}

\item{z_inf}{Maximum size of individuals [float]}

\item{n_0}{Number of individuals to initialize population with [integer]}

\item{init_dist}{The initial distribution to follow, three options 'Uniform', 'Random Uniform', 'Normal' \cr Uniform : Exactly uniformly distributed among all size classes \cr Random Uniform: Randomly uniformly distributed ~U[z_min,z_max] \cr Normal: Truncated normal distribution between 'z_min' and 'z_max' with mean = 'z_min'+'z_max'/2 and sd=1}
}
\value{
An initial size distribution [vector]
}
\description{
This creates an initial size distribution for indiviudals using either a uniform, random uniform, or random normal distribution
}
