% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSpawningProbs.R
\name{GenerateSpawningProbs}
\alias{GenerateSpawningProbs}
\title{Generate Spawning Probabilities}
\usage{
GenerateSpawningProbs(
  repro_start,
  repro_end,
  spawns_max_season,
  spawn_int,
  nReps = 1e+05,
  matReturn = F
)
}
\arguments{
\item{repro_start}{Onset of reproduction season in ordinal day (Jan 1 = 1 in non-leap year) [integer]}

\item{repro_end}{End of reproduction season in ordinal day [integer]}

\item{spawns_max_season}{Maximum number of spawns a female can have in a reproductive season [integer]}

\item{spawn_int}{Spawning interval or days between spawning events [integer]}

\item{nReps}{Number of individuals/reps used to generate simulation [integer]}

\item{matReturn}{True or False controls if a matrix is output. Default is FALSE, TRUE is often used for visualization.  [boolean]}
}
\value{
Daily spawning probability [vector] \cr Option: if(matReturn)=TRUE, output is a (nReps x (repro_end-repro_start)) binary matrix, where an entry of 1 in position i,j represents individual i spawning on day reproductive day j, useful for visualization
}
\description{
This function can be used to simulate daily spawning probability for a population of iteroparous or semelparous spawners based on a limited number of reproductive variables.
}
\note{
For iteroparous spawners, provide repro_start, repro_end, spawns_max_season, and spawn_int.

For semelparous spawners, provide repro_start, repro_end, and set spawns_max_season=1
}
\concept{Sub-models}
