% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateWinterSurvival.R
\name{GenerateWinterSurvival}
\alias{GenerateWinterSurvival}
\title{Generate Daily Winter Survival Probabilities}
\usage{
GenerateWinterSurvival(
  winterStartDate,
  winterEndDate,
  percentSurviving = 0.001
)
}
\arguments{
\item{winterStartDate}{Lethal concentration for 'lc_percent' [float]}

\item{winterEndDate}{Percentage lethality in population for 'lc_conc' [float]}

\item{percentSurviving}{The fraction of individuals below the winter size threshold that survive through winter. Default is 0.0001 [float]}
}
\value{
A time series (vector of length 365) of daily winter survival probabilities that are multiplied against background survival.  Outside of the winter start and end dates, the value is 1. [vector]
}
\description{
This function uses the length of the specified winter start and end dates to provide a modified daily winter surival probability
the creates a 'percentSurviving' probability of individuals below a specified winter cutoff to survive the entire winter
}
