% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthSurvivalKernel.R
\name{GrowthSurvivalKernel}
\alias{GrowthSurvivalKernel}
\title{Kernel Component -  Semelparous Adult Growth and Survival}
\usage{
GrowthSurvivalKernel(z1, z, bt, pars, date)
}
\arguments{
\item{z1}{Size at the end of the timestep, the size being transitioned to [float]}

\item{z}{Size at the beginning of the timestep [float]}

\item{bt}{Population biomass at the beginning of the timestep [float]}

\item{pars}{Data.frame containing the date-indexed parameters[data.frame]}

\item{date}{Ordinal day to reference proper 'pars' date-indexed parameters [integer]}
}
\value{
Size-dependent growth and survival kernel for adults
}
\description{
Integral projection models (Ellner et al., 2016) have the general structure of:
\deqn{n(z1,t+1) = \integral{ (P(z1,z) + R(z1,z))*n(z,t) dz}}
Here we define 'P' as 'GrowthSurvivalKernel' to be explicit, where it is the product of adult growth transition CDFs, \code{\link{Growth}}, with the size-dependent probability of survival, \code{\link{Survival}}.
\deqn{IteroparousGrowthSurvivalKernel(z1,z,bt,pars,date) = Growth(z1,z,bt,pars,date)*Survival(z,pars,date)}
}
\note{
This function is provided as a standard to pass to \code{\link{SimulateModel}} for the 'growthSurvivalComponent' argument

The 'SemelparousGrowthSurvivalKernel' is distinguished from the 'IteroparousSurvivalKernel' by in the inclusion of a (1-Spawning(z)) factor in the product for 'SemelparousGrowthSurvivalKernel' to account for post-spawning mortality.
}
\seealso{
Other Kernel Components: \code{\link{ReproductionKernel}}
}
\concept{Kernel Components}
