% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HatchlingGrowth.R
\name{HatchlingGrowth}
\alias{HatchlingGrowth}
\title{Kernel Functions - Hatchling Growth}
\usage{
HatchlingGrowth(z1, bt, pars, date)
}
\arguments{
\item{z1}{Size at the end of the timestep [float]}

\item{bt}{Population biomass at time t [float]}

\item{pars}{Data.frame containing the date-indexed parameters[data.frame]}

\item{date}{Ordinal day to reference proper 'pars' date-indexed parameters [integer]}
}
\value{
Growth CDF function for new hatchlings
}
\description{
Probabilistic Growth Transition Cumulative Distribution Function.
Uses vonBertalanffy to determine mean growth and a theoretical derivation
of the associated standard deviation.
}
\seealso{
Other Kernel Functions: \code{\link{Fecundity}},
  \code{\link{Spawning}}, \code{\link{Survival}}
}
\concept{Kernel Functions}
