% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-life_history_parameters_p.promelas.R
\docType{data}
\name{life_history_parameters_p.promelas}
\alias{life_history_parameters_p.promelas}
\title{Data: Fathead Minnow (Pimephales Promelas) Life History Parameters}
\format{data.frame}
\usage{
data(life_history_parameters_p.promelas)
}
\description{
Data: Fathead Minnow (Pimephales Promelas) Life History Parameters
}
\keyword{datasets}
