% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-parameters_master.R
\docType{data}
\name{parameters_master}
\alias{parameters_master}
\title{Data: Fish Toxicity Translator Master Parameter List}
\format{data.frame}
\usage{
data(life_history_template_master)
}
\description{
Data: Fish Toxicity Translator Master Parameter List
}
\keyword{datasets}
