% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalFunctions-Plotting.R
\name{PlotGrowth}
\alias{PlotGrowth}
\alias{PlotSurvival}
\alias{PlotReproduction}
\alias{PlotSpawningProbs}
\alias{PlotSurvivalDecrements}
\alias{PlotGrowthPercents}
\alias{PlotExposureConcentrations}
\title{Fish Toxicity Translator: Plotting Functions}
\usage{
PlotGrowth(pars, date, sizesToPlot = 100, bt = 0, plotSDevs = T)

PlotSurvival(pars, date, sizesToPlot = 100)

PlotReproduction(pars, date, sizesToPlot = 100)

PlotSpawningProbs(pars)

PlotSurvivalDecrements(pars)

PlotGrowthPercents(pars)

PlotExposureConcentrations(pars)
}
\arguments{
\item{pars}{A date indexed FishToxTranslator parameter data.frame() or list of data.frames}

\item{date}{Ordinal date used to access date specific FishToxTranslator parameters}

\item{sizesToPlot}{Number size classes/divisions to plot between the lower and upper size limits. Default=100.}

\item{bt}{For [PlotGrowth], Population biomass for date plotted.  These values are produced as an output from SimulateModel(), but can be input for explore different density dependent growth.  Default value is 0 to represent density independence.}

\item{plotSDevs}{For [PlotGrowth], T/F for including standard deviations of daily growth increments in the plot}
}
\value{
Returns a plot for visualization of FishToxTranslator model outputs and functions
}
\description{
This group of functions are used for plotting various components of the FishToxTranslator model
}
\section{Functions}{
\itemize{
\item \code{PlotGrowth}: This function plots the daily growth function by showing mean daily growth increment and standard deviation by size

\item \code{PlotSurvival}: This function plots the daily survival probability by size

\item \code{PlotReproduction}: This function plots the daily size-dependent reproductive output of females per female per day

\item \code{PlotSpawningProbs}: This function plots daily spawning probability

\item \code{PlotSurvivalDecrements}: This function plots daily survival decrements

\item \code{PlotGrowthPercents}: This function plots daily growth percents

\item \code{PlotExposureConcentrations}: This function plots daily exposure concentrations
}}

