% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproductionKernel.R
\name{ReproductionKernel}
\alias{ReproductionKernel}
\title{Kernel Component - Reproduction Kernel}
\usage{
ReproductionKernel(z1, z, bt, pars, date)
}
\arguments{
\item{z1}{Size at the end of the timestep, the size being transitioned to [float]}

\item{z}{Size at the beginning of the timestep [float]}

\item{bt}{Population biomass at the beginning of the timestep [float]}

\item{pars}{Data.frame containing the date-indexed parameters[data.frame]}

\item{date}{Ordinal day to reference proper 'pars' date-indexed parameters [integer]}
}
\value{
Size-dependent reproduction kernel
}
\description{
Integral projection models (Ellner et al., 2016) have the general structure of:
\deqn{n(z1,t+1) = \integral{ (P(z1,z) + R(z1,z))*n(z,t) dz}}
Here we define 'R' as 'ReproductionKernel' to b explicit, where it is the product size-dependent reproduction probability, \code{\link{Spawning}}, Size-dependent hatchlings per spawn, \code{\link{Fecundity}}, hatchling growth-transition CDF, \code{\link{HatchlingGrowth}}, and size-dependent survival, \code{\link{Survival}}.
\deqn{ReproductionKernel(z1,z,bt,pars,date) = sex_ratio * Spawning(z,pars,date) * Fecundity(z,pars,date)*HatchlingGrowth(z1,bt,pars,date)*Survival(z_hatch,pars,date)}
}
\note{
This function is provided as a standard to pass to \code{\link{SimulateModel}} for the 'reproductionComponent' argument
}
\seealso{
Other Kernel Components: \code{\link{GrowthSurvivalKernel}}
}
\concept{Kernel Components}
