% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spawning.R
\name{Spawning}
\alias{Spawning}
\title{Kernel Functions - Spawning Probability}
\usage{
Spawning(z, pars, date)
}
\arguments{
\item{z}{Size at the beginning of the timestep [float]}

\item{pars}{Data.frame containing the date-indexed parameters [data.frame]}

\item{date}{Ordinal day to reference proper 'pars' date-indexed parameters [integer]}
}
\value{
Probability of reproducing [binary]
}
\description{
Probability of reproducing, once an individual reaches 'z_repro' they reproduce with probability 'p_spawn'
}
\seealso{
Other Kernel Functions: 
\code{\link{Fecundity}()},
\code{\link{HatchlingGrowth}()},
\code{\link{Survival}()}
}
\concept{Kernel Functions}
