% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-species_library.R
\docType{data}
\name{species_library}
\alias{species_library}
\title{Data: Species Library}
\format{data.frame}
\usage{
data(parameters_p.promelas)
}
\description{
Data: Species Library
}
\keyword{datasets}
