% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{Survival}
\alias{Survival}
\title{Kernel Functions - Survival}
\usage{
Survival(z, pars, date)
}
\arguments{
\item{z}{Size at the beginning of the timestep [float]}

\item{pars}{Data.frame containing the date-indexed parameters[data.frame]}

\item{date}{Ordinal day to reference proper 'pars' date-indexed parameters [integer]}
}
\value{
A probabliity of survival [float]
}
\description{
Baseline Survival:
Logistic size-dependent daily survival probability for adults.
}
\details{
Winter Survival:
Logistic size-dependent daily survival probability for adults with a minimum cutoff size.

Exposure Survival:
A survival decrement can be specified to alter the daily survival probability for all individuals in the population.
}
\seealso{
Other Kernel Functions: \code{\link{Fecundity}},
  \code{\link{HatchlingGrowth}}, \code{\link{Spawning}}
}
\concept{Kernel Functions}
