% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalFunctions-DateTime.R
\name{TimeExt}
\alias{TimeExt}
\alias{OrdinalDate}
\alias{DateDuration}
\alias{SetWinterDates}
\alias{SetExposureDates}
\title{Time and Date Functions}
\usage{
TimeExt(sec = F)

OrdinalDate(date)

DateDuration(startDate, endDate)

SetWinterDates(winterStartDate, winterEndDate)

SetExposureDates(growth_percents)
}
\arguments{
\item{sec}{TRUE/FALSE to include seconds in time output}

\item{date}{Date in format 'DateMonthYear', ex. '01Jan20'}

\item{startDate}{Beginning ordinal date}

\item{endDate}{Ending ordinal date}

\item{winterStartDate}{Starting date for the duration [Ordinal date integer]}

\item{winterEndDate}{Ending date for the duration [Ordindal date interger]}

\item{growth_percents}{The daily Kappa growth percent decreases from baseline
return A vector of length length(growth_percents) with TRUE for each day where growth percent differs from 1}
}
\value{
A vector of length 365 where dates within 'startDate' and 'endDate' are TRUE all others are FALSE
}
\description{
These a group of functions that support other functions within the model
}
\section{Functions}{
\itemize{
\item \code{TimeExt}: Modified system time function formatted for use in naming files

\item \code{OrdinalDate}: Ordinal day function turns a date string into an integer

\item \code{DateDuration}: Ordinal day duration function, takes to ordinal dates and provides all ordinal dates in between them as a vector

\item \code{SetWinterDates}: This function is used to create a vector of TRUE/FALSE values where each day in the range specified is TRUE and those not within the duration are FALSE - used for winter dates here

\item \code{SetExposureDates}: This function takes a vector of growth percents from exposure and creates a boolean for each day that a growth percent exists.  Returns vector same length as input vector.
}}

